+++ backend/config/checker.js
@@ -1,5 +1,26 @@
 const database= require('./database.config');
 
+
+async function emailCheck (email){
+    let co;
+
+    try {
+        co = await database.getConnection();
+        const query = `SELECT email FROM project.project WHERE email=?`;
+        const values = [email];
+
+        const [result] = await co.execute(query,values);
+        co.release();
+        if (result.length>0) {
+            return true
+        }else {
+            return false
+        }
+    } catch (error) {
+        return error;
+    }
+}
+
 async function nicknameCheck (nickname){
     let co;
 
@@ -21,4 +42,4 @@ async function nicknameCheck (nickname){
 }
 
 
-module.exports = nicknameCheck;
+module.exports = {emailCheck,nicknameCheck};
