+++ backend/routes/users.js
@@ -1,14 +1,33 @@
 var express = require('express');
 var router = express.Router();
-const signup = require('./signup');
+const isCheck = require('./isCheck');
+const connect = require('../config/connection');
+
 
 const { swaggerUi, specs } = require('../swagger/swagger');
 
 
-router.post('/signup',  function(req, res, next) {
+router.post('/signup', function (req, res, next) {
   try {
     const { email, password, nickname } = req.body;
-    signup(email,password,nickname,res);
+
+    isCheck(email, password, nickname)
+      .then((result) => {
+        if (result.emcheck) {
+          return res.status(201).json({
+            message: "중복되는 정보가 존재합니다."
+          })
+        } else if (result.nicheck) {
+          return res.status(202).json({
+            message: "이미 있는 닉네임"
+          })
+        } else {
+          connect(email, password, nickname)
+          return res.status(200).json({
+            message: "완료"
+          })
+        }
+      })
   } catch (error) {
     return res.status(500).json({
       message: 'internal server error'
@@ -43,7 +62,7 @@ router.post('/signup',  function(req, res, next) {
  */
 
 
-router.post('/login',  function(req, res, next) {
+router.post('/login', function (req, res, next) {
   try {
     let { email, password } = req.body;
     console.log(email, password)
